% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rural_dz_lookup.R
\docType{data}
\name{rural_dz_lookup}
\alias{rural_dz_lookup}
\title{The Data Zone to SGUR lookup shows the relationship between Data Zones 2011 population weighted centroids and SGUR 2016.#'}
\format{Data frame with 6976 rows and 5 columns,
\describe{
  \item{DZ_CODE}{Census Output Area code}
  \item{UR2FOLD}{2-fold Urban Rural Classification}
  \item{UR3FOLD}{3-fold Urban Rural Classification}
  \item{UR6FOLD}{6-fold Urban Rural Classification}
  \item{UR8FOLD}{8-fold Urban Rural Classification}
}}
\source{
\url{https://www2.gov.scot/Resource/0054/00544928.csv}
}
\usage{
rural_dz_lookup
}
\description{
The Data Zone to SGUR lookup shows the relationship between Data Zones 2011 population weighted centroids and SGUR 2016.#'
}
\keyword{datasets}
