% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rural_oa_lookup.R
\docType{data}
\name{rural_oa_lookup}
\alias{rural_oa_lookup}
\title{The Output Area to SGUR lookup shows the relationship between the Census Output Area 2011 population weighted centroids and SGUR 2016.}
\format{Data frame with 46351 rows and 5 columns,
\describe{
  \item{OUTPUTAREA}{Census Output Area code}
  \item{UR2FOLD}{2-fold Urban Rural Classification}
  \item{UR3FOLD}{3-fold Urban Rural Classification}
  \item{UR6FOLD}{6-fold Urban Rural Classification}
  \item{UR8FOLD}{8-fold Urban Rural Classification}
}}
\source{
\url{https://www2.gov.scot/Resource/0054/00544929.csv}
}
\usage{
rural_oa_lookup
}
\description{
The Output Area to SGUR lookup shows the relationship between the Census Output Area 2011 population weighted centroids and SGUR 2016.
}
\keyword{datasets}
