% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rural_shapefile}
\alias{rural_shapefile}
\title{Shapefile of Rurality Classification}
\format{Stuff}
\source{
\url{}
}
\usage{
rural_shapefile
}
\description{
Shapefile of Rurality Classification
}
\keyword{datasets}
