% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ur6fold.R
\docType{data}
\name{ur6fold}
\alias{ur6fold}
\title{Six Fold Scottish Government Urban Rural Classification}
\format{A data frame with 6 rows and 3 columns,
\describe{
  \item{URCLASS}{6-Fold Urban Rural Class}
  \item{URNAME}{6-Fold Urban Rural Class name}
  \item{URDESC}{6-Fold Urban Rural Class description}
}}
\source{
\url{https://www2.gov.scot/Resource/0054/00544932.csv}
}
\usage{
ur6fold
}
\description{
Six Fold Scottish Government Urban Rural Classification
}
\keyword{datasets}
