% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ur8fold.R
\docType{data}
\name{ur8fold}
\alias{ur8fold}
\title{Eight Fold Scottish Government Urban Rural Classification}
\format{A data frame with 8 rows and 3 columns,
\describe{
  \item{URCLASS}{8-Fold Urban Rural Class}
  \item{URNAME}{8-Fold Urban Rural Class name}
  \item{URDESC}{8-Fold Urban Rural Class description}
}}
\source{
\url{https://www2.gov.scot/Resource/0054/00544933.csv}
}
\usage{
ur8fold
}
\description{
Eight Fold Scottish Government Urban Rural Classification
}
\keyword{datasets}
